+++ src/parsing_duckdb.jl
@@ -25,36 +25,36 @@ function expr_to_sql_duckdb(expr, sq; from_summarize::Bool)
                 return :(AVG($a))
             else
                 window_clause = construct_window_clause(sq)
-                return  "AVG($(string(a))) $(window_clause)"
+                return  "***AVG($(string(a))) $(window_clause)***"
             end
         elseif @capture(x, minimum(a_))
             if from_summarize
                 return :(MIN($a))
             else
                 window_clause = construct_window_clause(sq)
-                return  "MIN($(string(a))) $(window_clause)"
+                return  "***MIN($(string(a))) $(window_clause)***"
             end
         elseif @capture(x, maximum(a_))
             if from_summarize
                 return :(MAX($a))
             else
                 window_clause = construct_window_clause(sq)
-                return  "MAX($(string(a))) $(window_clause)"
+                return  "***MAX($(string(a))) $(window_clause)***"
             end
         elseif @capture(x, sum(a_))
             if from_summarize
                 return :(SUM($a))
             else
                 window_clause = construct_window_clause(sq)
-                return  "SUM($(string(a))) $(window_clause)"
+             #   println(sum_expr * window_clause)
+                return  "***SUM($(string(a))) $(window_clause)***"
             end
         elseif @capture(x, cumsum(a_))
             if from_summarize
                 error("cumsum is only available through a windowed @mutate") # COV_EXCL_LINE
             else
-               # sq.windowFrame = "ROWS UNBOUNDED PRECEDING "
                window_clause = construct_window_clause(sq, from_cumsum = true)
-               return  "SUM($(string(a))) $(window_clause)"
+               return  "***SUM($(string(a))) $(window_clause)***"
             end
         #stats agg
         elseif @capture(x, std(a_))