+++ NEWS.md
@@ -6,7 +6,8 @@
 - Improvements to CTE generation 
 - add kwarg `overwrite = false` to `copy_to` to default table copying to not replace exisiting tables with the name.
 - separate `@summary` into its own macro for collecting summary statistics (max, min, q1, q2, q3, avg, std, count, unique) from a table or file 
-
+- `@count` now produces column `n` not `count`
+- using sum in such a way will no longer break the SQL query: `@mutate(total = n / sum(n), _by = iv_abx)` 
 ## v0.8.0 - 2025-03-24
 - adds `@transmute`
 - adds `@separate` and `@unite`